// crates/app-server-protocol/schema/typescript/types/AssetVersionResult.ts
// Generated by `ash-app-server-protocol generate_protocol`. Do not edit.
import type { AssetImageType } from './AssetImageType.js';

export type AssetVersionResult = { assetId: string, versionId: string, name: string, source: string, sha256: string, mediaType: AssetImageType, size: number, width: number, height: number, };